+++ crates/astria-core/src/protocol/genesis/v1.rs
@@ -28,6 +28,7 @@ use crate::{
         transaction::v1::action::{
             BridgeLock,
             BridgeSudoChange,
+            BridgeTransfer,
             BridgeUnlock,
             FeeAssetChange,
             FeeChange,
@@ -36,6 +37,7 @@ use crate::{
             Ics20Withdrawal,
             InitBridgeAccount,
             PriceFeed,
+            RecoverIbcClient,
             RollupDataSubmission,
             SudoAddressChange,
             Transfer,
@@ -750,6 +752,7 @@ pub struct GenesisFees {
     pub init_bridge_account: Option<FeeComponents<InitBridgeAccount>>,
     pub bridge_lock: Option<FeeComponents<BridgeLock>>,
     pub bridge_unlock: Option<FeeComponents<BridgeUnlock>>,
+    pub bridge_transfer: Option<FeeComponents<BridgeTransfer>>,
     pub bridge_sudo_change: Option<FeeComponents<BridgeSudoChange>>,
     pub ibc_relay: Option<FeeComponents<IbcRelay>>,
     pub validator_update: Option<FeeComponents<ValidatorUpdate>>,
@@ -758,6 +761,7 @@ pub struct GenesisFees {
     pub ibc_relayer_change: Option<FeeComponents<IbcRelayerChange>>,
     pub sudo_address_change: Option<FeeComponents<SudoAddressChange>>,
     pub ibc_sudo_change: Option<FeeComponents<IbcSudoChange>>,
+    pub recover_ibc_client: Option<FeeComponents<RecoverIbcClient>>,
     pub price_feed: Option<FeeComponents<PriceFeed>>,
 }
 
@@ -777,6 +781,7 @@ impl Protobuf for GenesisFees {
             init_bridge_account,
             bridge_lock,
             bridge_unlock,
+            bridge_transfer,
             bridge_sudo_change,
             ibc_relay,
             validator_update,
@@ -785,6 +790,7 @@ impl Protobuf for GenesisFees {
             ibc_relayer_change,
             sudo_address_change,
             ibc_sudo_change,
+            recover_ibc_client,
             price_feed,
         } = raw;
         let rollup_data_submission = rollup_data_submission
@@ -822,6 +828,12 @@ impl Protobuf for GenesisFees {
             .map(FeeComponents::<BridgeUnlock>::try_from_raw)
             .transpose()
             .map_err(|e| FeesError::fee_components("bridge_unlock", e))?;
+
+        let bridge_transfer = bridge_transfer
+            .clone()
+            .map(FeeComponents::<BridgeTransfer>::try_from_raw)
+            .transpose()
+            .map_err(|e| FeesError::fee_components("bridge_transfer", e))?;
 
         let bridge_sudo_change = bridge_sudo_change
             .clone()
@@ -872,6 +884,12 @@ impl Protobuf for GenesisFees {
             .transpose()
             .map_err(|e| FeesError::fee_components("ibc_sudo_change", e))?;
 
+        let recover_ibc_client = recover_ibc_client
+            .clone()
+            .map(FeeComponents::<RecoverIbcClient>::try_from_raw)
+            .transpose()
+            .map_err(|e| FeesError::fee_components("recover_ibc_client", e))?;
+
         let price_feed = price_feed
             .clone()
             .map(FeeComponents::<PriceFeed>::try_from_raw)
@@ -885,6 +903,7 @@ impl Protobuf for GenesisFees {
             init_bridge_account,
             bridge_lock,
             bridge_unlock,
+            bridge_transfer,
             bridge_sudo_change,
             ibc_relay,
             validator_update,
@@ -893,6 +912,7 @@ impl Protobuf for GenesisFees {
             ibc_relayer_change,
             sudo_address_change,
             ibc_sudo_change,
+            recover_ibc_client,
             price_feed,
         })
     }
@@ -905,6 +925,7 @@ impl Protobuf for GenesisFees {
             init_bridge_account,
             bridge_lock,
             bridge_unlock,
+            bridge_transfer,
             bridge_sudo_change,
             ibc_relay,
             validator_update,
@@ -913,6 +934,7 @@ impl Protobuf for GenesisFees {
             ibc_relayer_change,
             sudo_address_change,
             ibc_sudo_change,
+            recover_ibc_client,
             price_feed,
         } = self;
         Self::Raw {
@@ -925,6 +947,8 @@ impl Protobuf for GenesisFees {
                 .map(|act| FeeComponents::<InitBridgeAccount>::to_raw(&act)),
             bridge_lock: bridge_lock.map(|act| FeeComponents::<BridgeLock>::to_raw(&act)),
             bridge_unlock: bridge_unlock.map(|act| FeeComponents::<BridgeUnlock>::to_raw(&act)),
+            bridge_transfer: bridge_transfer
+                .map(|act| FeeComponents::<BridgeTransfer>::to_raw(&act)),
             bridge_sudo_change: bridge_sudo_change
                 .map(|act| FeeComponents::<BridgeSudoChange>::to_raw(&act)),
             ibc_relay: ibc_relay.map(|act| FeeComponents::<IbcRelay>::to_raw(&act)),
@@ -939,6 +963,8 @@ impl Protobuf for GenesisFees {
                 .map(|act| FeeComponents::<SudoAddressChange>::to_raw(&act)),
             ibc_sudo_change: ibc_sudo_change
                 .map(|act| FeeComponents::<IbcSudoChange>::to_raw(&act)),
+            recover_ibc_client: recover_ibc_client
+                .map(|act| FeeComponents::<RecoverIbcClient>::to_raw(&act)),
             price_feed: price_feed.map(|act| FeeComponents::<PriceFeed>::to_raw(&act)),
         }
     }
@@ -1075,6 +1101,7 @@ mod tests {
             init_bridge_account: Some(FeeComponents::<InitBridgeAccount>::new(48, 0).to_raw()),
             bridge_lock: Some(FeeComponents::<BridgeLock>::new(12, 1).to_raw()),
             bridge_unlock: Some(FeeComponents::<BridgeUnlock>::new(12, 0).to_raw()),
+            bridge_transfer: Some(FeeComponents::<BridgeTransfer>::new(24, 0).to_raw()),
             bridge_sudo_change: Some(FeeComponents::<BridgeSudoChange>::new(24, 0).to_raw()),
             ics20_withdrawal: Some(FeeComponents::<Ics20Withdrawal>::new(24, 0).to_raw()),
             ibc_relay: Some(FeeComponents::<IbcRelay>::new(0, 0).to_raw()),
@@ -1084,6 +1111,7 @@ mod tests {
             ibc_relayer_change: Some(FeeComponents::<IbcRelayerChange>::new(0, 0).to_raw()),
             sudo_address_change: Some(FeeComponents::<SudoAddressChange>::new(0, 0).to_raw()),
             ibc_sudo_change: Some(FeeComponents::<IbcSudoChange>::new(0, 0).to_raw()),
+            recover_ibc_client: Some(FeeComponents::<RecoverIbcClient>::new(0, 0).to_raw()),
             price_feed: Some(FeeComponents::<PriceFeed>::new(0, 0).to_raw()),
         }
     }