+++ crates/astria-sequencer-utils/src/genesis_example.rs
@@ -39,6 +39,7 @@ use astria_core::{
         transaction::v1::action::{
             BridgeLock,
             BridgeSudoChange,
+            BridgeTransfer,
             BridgeUnlock,
             FeeAssetChange,
             FeeChange,
@@ -47,6 +48,7 @@ use astria_core::{
             Ics20Withdrawal,
             InitBridgeAccount,
             PriceFeed,
+            RecoverIbcClient,
             RollupDataSubmission,
             SudoAddressChange,
             Transfer,
@@ -239,6 +241,7 @@ fn proto_genesis_state() -> astria_core::generated::astria::protocol::genesis::v
             init_bridge_account: Some(FeeComponents::<InitBridgeAccount>::new(48, 0).to_raw()),
             bridge_lock: Some(FeeComponents::<BridgeLock>::new(12, 1).to_raw()),
             bridge_unlock: Some(FeeComponents::<BridgeUnlock>::new(12, 0).to_raw()),
+            bridge_transfer: Some(FeeComponents::<BridgeTransfer>::new(24, 0).to_raw()),
             bridge_sudo_change: Some(FeeComponents::<BridgeSudoChange>::new(24, 0).to_raw()),
             ics20_withdrawal: Some(FeeComponents::<Ics20Withdrawal>::new(24, 0).to_raw()),
             ibc_relay: Some(FeeComponents::<IbcRelay>::new(0, 0).to_raw()),
@@ -248,6 +251,7 @@ fn proto_genesis_state() -> astria_core::generated::astria::protocol::genesis::v
             ibc_relayer_change: Some(FeeComponents::<IbcRelayerChange>::new(0, 0).to_raw()),
             sudo_address_change: Some(FeeComponents::<SudoAddressChange>::new(0, 0).to_raw()),
             ibc_sudo_change: Some(FeeComponents::<IbcSudoChange>::new(0, 0).to_raw()),
+            recover_ibc_client: Some(FeeComponents::<RecoverIbcClient>::new(0, 0).to_raw()),
             price_feed: Some(FeeComponents::<PriceFeed>::new(0, 0).to_raw()),
         }),
     }