+++ crates/astria-composer/CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Bump MSRV to 1.83.0 [#1857](https://github.com/astriaorg/astria/pull/1857).
 - Bump penumbra dependencies [#1740](https://github.com/astriaorg/astria/pull/1740).
+- Propagate errors [#1838](https://github.com/astriaorg/astria/pull/1838).
 
 ## [1.0.0-rc.2] - 2024-10-23